refactor: HealthCheck 응답 데이터 위치 변경 및 테스트 수정
