+++ src/main/kotlin/me/study/kotlin/controller/HealthController.kt
@@ -1,9 +1,12 @@
 package me.study.kotlin.controller
 
+import me.study.kotlin.dto.HealthResponse
 import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 
 @RestController
+@RequestMapping("/api")
 class HealthController {
 
     @GetMapping("/health")
