+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
+server:
+  port: 8080
+
 spring:
   application:
     name: kotlin
-  mvc:
-    servlet:
-      path: /api
